Reorder exports and tighten resource handling
Refactor module exports and clean up numerous small issues across the codebase. Reorganized many __all__ lists (primr, agentic, ai, api, config, core, data, subagents) to consistently export types and APIs; updated pyproject ruff ignores. Improve resource handling and simplify logic: combine nested with statements for DB access in ContentCache, use contextlib.suppress for errno-safe cleanup, replace loops with any/all comprehensions, remove redundant pass statements, and consolidate contextmanager usage. Fix small bugs and clarity changes (e.g. startswith tuple grouping, rename response->user_choice, use list spread for prepending consolidated context file). Overall these changes tighten locking/cleanup, reduce boilerplate, and standardize exports.

+++ src/primr/qa/issue_classifier.py
@@ -130,7 +130,7 @@ class IssueClassifier:
         severity_penalty = severity_impact * 20  # Up to 20 point penalty
 
         final_score = max(0, min(100, base_score - severity_penalty))
-        return int(round(final_score))
+        return round(final_score)
 
     def _calculate_issues_score(self, issues: list[ClassifiedIssue]) -> int:
         """
